EXPERIMENTAL Make Vrapper receive commands depending on mode
This experiment does two things:

1. It adds a "Vrapper Input" command to Eclipse's Keys preferences page
to which you can assign a key. It also adds a number of contexts which
can be selected on that same page via the "When" field to make such key
binds conditional. These key binds will then overrule any other keys as
long as the context in the "When" field is active. This way a user
doesn't need to unbind any Eclipse keys and can toggle between vanilla
Eclipse and Vrapper use.

2. It adds a bit of code and a bunch of Eclipse Core Expressions in
plugin.xml to intercept the "...lineStart" and "...lineEnd" commands
when the command line context is active. This way users don't even need
to mess with their Keys preferences as they would likely still bind it
to the HOME and END keys.

The bad side is that this experiment is playing a game of walls and
ladders; it adds a bunch of useless contexts and expression references
just to make Vrapper look more important than any of the editor plugins.
The result is that future Eclipse updates might break it when the
precedence rules would be changed or when some other editor manages to
look as important as Vrapper (read: it has just as many or more stacked
contexts or expression references).