EXPERIMENTAL: delegate to EclipseCommandHandler, one instance per editor
The EclipseCommandHandler is instantiated for each separate editor. This
should allow it to hold more editor-specific state while the command
detection is decoupled from the logic.

Note that each top-level editor has a special Eclipse command service
which would only be triggered when that top-level editor is active. This
one is not used because multi-page editors have several
InputInterceptors, and thus several command listeners would get hooked
which meant all nested editors would be triggered at the same time.

The workbench-wide listener is also less prone to bugs as certain
versions of Eclipse 3.8/4.2 would not call the postExecuteFailure method
when a command could not be executed.