EXPERIMENTAL: Move textobject and motion command ids to plugin.xml
Refactoring this as an extension point allows to configure these command
ids without a full rebuild, but it would also allow someone to create a
custom plugin which registers a whole bunch of commands for use in a
custom Eclipse IDE distribution.

Also make sure to record Eclipse commands for all known command types.