Add command and handler to force focus to current editor
Currently there are a number of situations in which Eclipse will stop
showing the caret and fail to accept input, for example when the user
closes the last document in a tab group created through vsplit.

There isn't much we can do about that, so instead we can add a new
Workbench command / handler combo which can be used instead of the
default Activate Editor command. Our own version will not only switch to
the currently active editor if a View is active; it will also look up
the StyledText widget of an associated Vrapper instance and "replay" the
keypress to that widget. This makes it ideal to bind Escape (or Ctrl+[
if you prefer) to this command so that it will both activate the editor
and reset the current mode (if it wasn't already in Normal mode).

The handler will remain active if Vrapper is disabled, though this
shouldn't be a big deal because most editors do not really use the
Escape key, and even then they should still receive the keypress thanks
to the replay attempt.

If the current editor cannot be wrapped by Vrapper then we will only
activate it and it should run its ownn activation / focus logic.

Fixes #435.