Implement an ISourceProvider which notifies Eclipse of Vrapper's state
The current mode name and the "isVrapperEnabled" state can then be
queried from Eclipse Core Expressions which is used to activate Eclipse
commands and handlers.

The mode will get updated each time the editor is activated again after
it lost focus or when the user is switching pages inside of a multi-page
editor. If the user switches to an unknown editor then the mode name
will be set to "(unknown mode)".

Note that the mode will not be unset when the focus of the editor is
lost. Any core expression which uses the mode should instead check that
the currently active part is of type IEditorPart.