Add custom Vrapper keybinding contexts, hook context activation expressions
Eclipse shortcuts are context-sensitive thanks to the "When:" dropdown
in the key binding dialog. This way a given key combination can mean
different things depending on which editor or view is active.

Vrapper users have been asking for a way to bind keys to Vrapper in such
a way that it can be immediately undone when Vrapper is disabled. Now
that we have the Vrapper ISourceProvider we can finally add a "When:"
context which is toggled by the Vrapper machinery.

We request Eclipse to activate our custom contexts when a given
Expression is true; this expression will decide the state of each
context depending on the Vrapper mode variables. Eclipse is then smart
enough to re-evaluate all those expressions each time the state changes
when your ISourceProvider notifies Eclipse.

Work for #108.