EXPERIMENTAL First attempt at porting EasyMotion to Vrapper
This uses the new label infrastructure.

Targetting is complete, but the available keys for the target grouping
algorithm still need to be made configurable.

The most important missing bit is that it currently uses pairs of Motion
instances for finding targets, which turns out to be a bad choice. These
Motions may use EditorAdapter.get/setPosition, and it can be clearly
seen that the editor's current top line gets shifted when targetting up
or down.

Also, character input mode is still missing. This makes f/F/t/T or
Sneak-like motions impossible for the moment.